fix: replaced old Detailed.js by new
