+++ src/components/PostVisualBacktest.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { graphql } from 'react-apollo'
 import { compose, withProps } from 'recompose'
 import LazyLoad from 'react-lazyload'
-import { HistoryPriceByTickerGQL } from './../pages/Detailed/DetailedGQL'
+import { HistoryPriceByTickerGQL } from '../pages/Detailed/gqlWrappers/DetailedGQL'
 import PercentChanges from './PercentChanges'
 import PostVisualBacktestChart from './PostVisualBacktestChart'
 import { binarySearchHistoryPriceIndex } from '../utils/utils'
