--- src/components/ProjectChart/ProjectChart.css
@@ -1,142 +1,0 @@
-.project-dp-chart {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-}
-
-.project-dp-chart .graph {
-  flex: 1;
-}
-
-@media all and (orientation: portrait) {
-  .project-dp-chart-mobile-fullscreen {
-    transform: rotate(-90deg);
-    transform-origin: left top;
-    width: 100vh;
-    overflow-x: hidden;
-    position: absolute;
-    top: 100%;
-    left: 0;
-    background: white;
-  }
-}
-
-.project-chart-loader {
-  padding-top: 10px;
-}
-
-.project-chart-body {
-  background: rgba(245, 249, 255, 0.6);
-  position: relative;
-  padding: 0 20px;
-  padding-top: 10px;
-}
-
-.project-chart__isLoading,
-.project-chart__isEmpty {
-  color: rgb(38, 43, 51);
-  font-size: 2em;
-  font-weight: 300;
-  position: absolute;
-  display: flex;
-  align-items: center;
-  height: 100%;
-  width: 100%;
-  justify-content: center;
-  user-select: none;
-  z-index: 9999;
-}
-
-.chart-datetime-settings {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.time-filter {
-  margin-right: 20px;
-}
-
-.selected-value {
-  font-size: 1em;
-  text-align: center;
-}
-
-.selected-value-datetime {
-  font-size: 1.4em;
-  font-weight: 300;
-}
-
-.selected-value-data {
-  display: flex;
-  flex-direction: column;
-  font-size: 1em;
-  font-weight: bold;
-  margin-right: 0.5em;
-  text-align: left;
-}
-
-@media (max-width: 1024px) {
-  .selected-value-datetime {
-    font-size: 1em;
-    font-weight: 300;
-  }
-
-  .selected-value-data {
-    display: flex;
-    flex-direction: column;
-    font-size: 0.8em;
-    font-weight: bold;
-    margin-right: 0;
-    text-align: right;
-  }
-}
-
-.time-filter,
-.currency-filter {
-  display: flex;
-  user-select: none;
-}
-
-.time-filter div,
-.currency-filter div,
-.chart-footer-filters div {
-  cursor: pointer;
-  min-width: 32px;
-  text-align: center;
-  border-radius: 4px;
-}
-
-.time-filter .activated,
-.currency-filter .activated,
-.chart-footer-filters .activated {
-  background: rgba(38, 43, 51, 0.8);
-  color: white;
-}
-
-.time-filter .disabled,
-.currency-filter .disabled,
-.chart-footer-filters .disabled {
-  color: #888;
-}
-
-.night-mode .time-filter,
-.night-mode .currency-filter,
-.night-mode .chart-footer-filters,
-.night-mode .time-filter .activated,
-.night-mode .currency-filter .activated,
-.night-mode .chart-footer-filters .activated {
-  color: #a0a0a0;
-}
-
-.night-mode .time-filter .activated,
-.night-mode .currency-filter .activated,
-.night-mode .chart-footer-filters .activated {
-  background-color: #252932;
-}
-
-.night-mode .time-filter .disabled,
-.night-mode .currency-filter .disabled,
-.night-mode .chart-footer-filters .disabled {
-  color: #595959;
-}
