--- src/components/ProjectChart/ProjectChartFooter.css
@@ -1,234 +1,0 @@
-.chart-footer {
-  border-top: 1px solid rgb(216, 216, 218);
-  display: flex;
-  flex-direction: column;
-}
-
-.toggleBtn {
-  display: flex;
-  position: relative;
-  padding: 2px 10px;
-  align-items: center;
-  justify-content: flex-end;
-  user-select: none;
-  cursor: pointer;
-  text-align: center;
-  border-radius: 4px;
-  box-shadow: 0 4px 6px rgba(50, 50, 92, 0.11), 0 1px 3px rgba(0, 0, 0, 0.075);
-}
-
-.toggleBtn div i {
-  margin-right: 0;
-}
-
-.toggleBtn .label {
-  margin-right: 5px !important;
-  font-size: 10px;
-}
-
-.toggleBtn .icon {
-  margin-bottom: 6px;
-  margin-left: 3px;
-}
-
-.toggleBtn.activated {
-  background: rgba(38, 43, 51, 0.8);
-  color: white;
-}
-
-.toggleBtn.disabled {
-  color: #888;
-}
-
-.toggleBtn .toggleBtn--loading {
-  color: transparent !important;
-  text-shadow: 0 0 3px rgba(0, 0, 0, 0.3) !important;
-}
-
-.toggleBtn .toggleBtn-loader {
-  position: absolute;
-  top: 50%;
-  left: 40%;
-  color: white;
-}
-
-.marketcapLabel {
-  color: rgb(52, 118, 153) !important;
-  background-color: rgb(52, 118, 153) !important;
-  border-color: rgb(52, 118, 153) !important;
-}
-
-.volumeLabel {
-  color: rgba(38, 43, 51, 0.25) !important;
-  background-color: rgba(38, 43, 51, 0.25) !important;
-  border-color: rgba(38, 43, 51, 0.25) !important;
-}
-
-.devActivityLabel {
-  color: rgb(96, 76, 141) !important;
-  background-color: rgb(96, 76, 141) !important;
-  border-color: rgb(96, 76, 141) !important;
-}
-
-.burnRateLabel {
-  color: rgb(252, 138, 23) !important;
-  background-color: rgb(252, 138, 23) !important;
-  border-color: rgb(252, 138, 23) !important;
-}
-
-.transactionVolumeLabel {
-  color: rgb(39, 166, 153) !important;
-  background-color: rgb(39, 166, 153) !important;
-  border-color: rgb(39, 166, 153) !important;
-}
-
-.twitterLabel {
-  color: rgb(16, 195, 245) !important;
-  background-color: rgb(16, 195, 245) !important;
-  border-color: rgb(16, 195, 245) !important;
-}
-
-.filter-category-body a {
-  margin-top: 12px;
-}
-
-.sentimentLabel {
-  color: #e23ab4 !important;
-  background-color: #e23ab4 !important;
-  border-color: #e23ab4 !important;
-}
-
-.ethSpentOverTimeLabel {
-  color: #c82f3f !important;
-  background-color: #c82f3f !important;
-  border-color: #c82f3f !important;
-}
-
-.chart-footer-filters p {
-  margin: 5px !important;
-}
-
-.filter-category {
-  display: flex;
-  padding: 16px;
-  margin: 15px 0;
-  border-right: 1px solid rgb(231, 231, 231);
-  user-select: none;
-  justify-content: space-between;
-  align-items: normal;
-}
-
-.filter-category:first-child {
-  width: 230px;
-}
-
-.night-mode .filter-category {
-  border-color: #2d3239;
-  border-radius: 0;
-}
-
-.filter-category:last-child {
-  border: none;
-}
-
-.filter-category-title {
-  margin: 0;
-  font-weight: 400;
-}
-
-.filter-category-body {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  max-width: 30%;
-}
-
-.filter-category-body .toggleBtn:last-child {
-  margin-right: 0;
-}
-
-.filter-category-blockchain .filter-category-body {
-  flex-direction: column;
-  justify-content: end;
-}
-
-.filter-category-blockchain .filter-category-title {
-  flex: 1;
-  text-align: left;
-}
-
-.blockchain-settings {
-  display: flex;
-  flex-direction: column;
-}
-
-.blockchain-settings .checkbox {
-  margin-bottom: 5px;
-}
-
-.blockchain-settings .checkbox:last-child {
-  margin-bottom: 0;
-}
-
-.premium-wall-button {
-  border: 2px solid gold;
-  margin-top: 5px;
-}
-
-.night-mode .chart-footer-filters .toggleBtn {
-  border: 1px solid #252932;
-}
-
-.night-mode .chart-footer-filters .premium-wall-button {
-  border-color: #a18825;
-  border-width: 2px;
-}
-
-@media (min-width: 1024px) {
-  .chart-footer {
-    display: flex;
-  }
-
-  .chart-footer-filters {
-    display: flex;
-    justify-content: center;
-  }
-
-  .filter-category {
-    display: block;
-    padding: 5px 20px;
-  }
-
-  .filter-category-title {
-    margin-bottom: 10px;
-    flex: 1;
-  }
-
-  .filter-category-body {
-    max-width: 100%;
-  }
-
-  .filter-category-blockchain {
-    justify-content: end;
-  }
-
-  .filter-category-blockchain .filter-category-title {
-    text-align: center;
-  }
-
-  .filter-category-blockchain .filter-category-body {
-    justify-content: end;
-  }
-
-  .toggleBtn {
-    justify-content: flex-start;
-  }
-
-  .filter-category-body .toggleBtn {
-    margin-top: 10px;
-  }
-
-  .filter-category-body .toggleBtn:first-child {
-    margin-top: 0;
-  }
-}
