--- src/components/ProjectChart/ProjectChartHeader.css
@@ -1,23 +1,0 @@
-.chart-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0;
-  min-height: 60px;
-  border-bottom: 1px solid rgb(216, 216, 218);
-}
-
-.chart-header-actions {
-  display: flex;
-  align-items: center;
-}
-
-.chart-header-actions .currency-filter {
-  margin-right: 20px;
-}
-
-@media (min-width: 1024px) {
-  .chart-header {
-    padding: 0 20px;
-  }
-}
