--- src/components/ProjectChart/ProjectChartMobile.css
@@ -1,28 +1,0 @@
-.detailed-page-mobile-settings-bar {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  z-index: 9999;
-  background-color: white;
-  box-shadow: 0 -1px 4px 0 rgba(0, 0, 0, 0.2);
-  padding: 6px;
-  display: flex;
-  justify-content: flex-end;
-}
-
-.project-dp-chart-mobile-fullscreen .chart-footer {
-  padding: 0 20px;
-}
-
-.ico-price-label {
-  display: flex;
-}
-
-.ico-price-legend {
-  width: 20px;
-  height: 1px;
-  margin-top: 10px;
-  margin-left: 10px;
-  border: 1px dashed black;
-}
