+++ src/pages/Detailed/Detailed.module.scss
@@ -2,3 +2,7 @@
   width: 100%;
   border: none;
 }
+
+.desktopCharts {
+  min-height: 428px;
+}
