+++ src/pages/Detailed/financialInfo/FinancialsBlock.css
