+++ src/pages/Detailed/generalInfo/GeneralInfoBlock.module.scss
