+++ src/pages/Detailed/gqlWrappers/DetailedGQL.js
