+++ src/pages/Detailed/gqlWrappers/withTimeseries.js
