+++ src/pages/Detailed/header/DetailedHeader.module.scss
