+++ src/pages/Detailed/mobile/MobileAssetChart.module.scss
