+++ src/pages/Detailed/mobile/MobileDetailedPage.js
@@ -3,30 +3,30 @@ import cx from 'classnames'
 import { compose } from 'recompose'
 import { graphql } from 'react-apollo'
 import { Label, Selector } from '@santiment-network/ui'
-import { DailyActiveAddressesGQL } from './DetailedGQL'
-import { TRANSACTION_VOLUME_QUERY } from '../../ducks/GetTimeSeries/queries/transaction_volume_query'
-import { SOCIAL_VOLUME_QUERY } from '../../ducks/GetTimeSeries/queries/social_volume_query'
-import { NEWS_QUERY } from '../../components/News/NewsGQL'
-import { calcPercentageChange, capitalizeStr } from './../../utils/utils'
+import { DailyActiveAddressesGQL } from '../gqlWrappers/DetailedGQL'
+import { TRANSACTION_VOLUME_QUERY } from '../../../ducks/GetTimeSeries/queries/transaction_volume_query'
+import { SOCIAL_VOLUME_QUERY } from '../../../ducks/GetTimeSeries/queries/social_volume_query'
+import { NEWS_QUERY } from '../../../components/News/NewsGQL'
+import { calcPercentageChange, capitalizeStr } from '../../../utils/utils'
 import {
   DAY,
   getTimeIntervalFromToday,
   getIntervalByTimeRange
-} from '../../utils/dates'
-import { formatNumber } from './../../utils/formatting'
-import MobileHeader from './../../components/MobileHeader/MobileHeader'
-import PercentChanges from './../../components/PercentChanges'
-import NewsSmall from '../../components/News/NewsSmall'
-import PageLoader from '../../components/Loader/PageLoader'
-import MobileMetricCard from '../../components/MobileMetricCard/MobileMetricCard'
-import GetAsset from './GetAsset'
-import GetTimeSeries from '../../ducks/GetTimeSeries/GetTimeSeries'
+} from '../../../utils/dates'
+import { formatNumber } from '../../../utils/formatting'
+import MobileHeader from '../../../components/MobileHeader/MobileHeader'
+import PercentChanges from '../../../components/PercentChanges'
+import NewsSmall from '../../../components/News/NewsSmall'
+import PageLoader from '../../../components/Loader/PageLoader'
+import MobileMetricCard from '../../../components/MobileMetricCard/MobileMetricCard'
+import GetAsset from '../gqlWrappers/GetAsset'
+import GetTimeSeries from '../../../ducks/GetTimeSeries/GetTimeSeries'
 import MobileAssetChart from './MobileAssetChart'
-import ShowIf from '../../components/ShowIf'
-import GetWatchlists from '../../ducks/Watchlists/GetWatchlists'
-import WatchlistsPopup from '../../components/WatchlistPopup/WatchlistsPopup'
-import { mergeTimeseriesByKey } from '../../utils/utils'
-import { addRecentAssets } from '../../utils/recent'
+import ShowIf from '../../../components/ShowIf'
+import GetWatchlists from '../../../ducks/Watchlists/GetWatchlists'
+import WatchlistsPopup from '../../../components/WatchlistPopup/WatchlistsPopup'
+import { mergeTimeseriesByKey } from '../../../utils/utils'
+import { addRecentAssets } from '../../../utils/recent'
 import styles from './MobileDetailedPage.module.scss'
 
 const MobileDetailedPage = props => {
