+++ src/pages/Detailed/mobile/MobileDetailedPage.module.scss
