+++ src/pages/Detailed/transactionsInfo/DetailedTransactionsTable.css
